Hull_go, star_end free fly, hidden FAI
Creation of the hull_go function to call the ConvexHull algorithm only once. Within this function, the track is divided into 5 parts to find more points (it will be necessary to determine how to adjust this number based on the required precision or speed).

Function to calculate the start and end points of a free distance track.

Search for hidden FAI triangles.